% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade.pa.R
\name{clade.pa}
\alias{clade.pa}
\title{Takes a vc.clade obect, returns a stack of PA rasters, a table with presences
for all species, and a revised clade with presence data.  Bias raster is multiplied by
suitability to get sample probability, and if npres isn't NA only that many presences
will be returned.}
\usage{
clade.pa(
  x,
  sample.source = "suitab.raster",
  bias.raster = NA,
  npres = NA,
  method = "auto",
  ...
)
}
\arguments{
\item{x}{A vc.clade object to generate pa data from}

\item{sample.source}{Either "suitab.raster" or "actual.range".  If the latter is provided the species all need to have a raster provided containing their ranges.}

\item{bias.raster}{A raster depicting probability of sampling in each grid cell for the entire region.}

\item{npres}{Number of presence points per species.  If not provided, it will just return the number of grid cells that are deemed "present" by virtualspecies converToPA function.}

\item{method}{If "auto", the function tries to automatically pick good values for virtualspecies' convertToPA function based on the desired number of presences.  Otherwise just passes the "method" argument into the PA.method of convertToPA.}
}
\description{
Takes a vc.clade obect, returns a stack of PA rasters, a table with presences
for all species, and a revised clade with presence data.  Bias raster is multiplied by
suitability to get sample probability, and if npres isn't NA only that many presences
will be returned.
}
