% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.clade.R
\name{evolve.clade}
\alias{evolve.clade}
\title{Evolves a clade of virtual species from a single virtualspecies object.}
\usage{
evolve.clade(
  root.species,
  ntaxa,
  env,
  rate = 0.2,
  min.sd = 0.5,
  tip.prefix = "species",
  max.tries = 10,
  ...
)
}
\arguments{
\item{root.species}{A virtualspecies object to be used as the common ancestor of all simulated species.  Note that this currently has to be one that was generated with approach = "response".}

\item{ntaxa}{Number of tip taxa desired.}

\item{env}{A stack of environmental rasters.}

\item{rate}{The Brownian rate parameter controlling the rate of niche evolution.  Since different niche axes will have different parameter values, this rate is expressed as as a proportion of the value of each parameter at the root node. So if rate = 0.5 and the value of a parameter for the root species is 4, the sigma parameter for ape's rTraitCont simulation will be 0.5 * 4 = 2.}

\item{min.sd}{The minimum standard deviation for each parameter, as a proportion of the parameter value at the root.  Essentially this sets the minimum possible niche width.}

\item{tip.prefix}{The prefix to use for naming species.}
}
\value{
A list containing the root species, the tip species, and a phylogeny.  Also contains the arguments used to run the simulation.
}
\description{
Evolves a clade of virtual species from a single virtualspecies object.
}
\examples{
my.root <- generateRandomSp(euro.worldclim[[c(1,12)]])
my.clade <- evolve.clade(my.root, 10, euro.worldclim)

}
