% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.plots.R
\name{parameter.plots}
\alias{parameter.plots}
\title{Plots parameter values for a set of species in a virtualclade object.  Returns both histograms of the distribution of parameter values and a set of trees with parameter values mapped on them.}
\usage{
parameter.plots(clade)
}
\arguments{
\item{clade}{A virtualclade object}
}
\value{
A list containing the plotting data frame and a series of plots.
}
\description{
Plots parameter values for a set of species in a virtualclade object.  Returns both histograms of the distribution of parameter values and a set of trees with parameter values mapped on them.
}
\examples{
my.root <- generateRandomSp(euro.worldclim[[c(1,12)]])
my.clade <- evolve.clade(my.root, 10, euro.worldclim)
parameter.plots(my.clade)

}
