% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.plots.R
\name{response.plots}
\alias{response.plots}
\title{Plots responses to all environmental variables for a set of species in a virtualclade object}
\usage{
response.plots(clade)
}
\arguments{
\item{clade}{A virtualclade object}
}
\value{
A list containing the plotting data frame and a ggplot object showing the response of each species to the environmental gradient.
}
\description{
Plots responses to all environmental variables for a set of species in a virtualclade object
}
\examples{
my.root <- generateRandomSp(euro.worldclim[[c(1,12)]])
my.clade <- evolve.clade(my.root, 10, euro.worldclim)
response.plots(my.clade)

}
