% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpack.species.R
\name{unpack.species}
\alias{unpack.species}
\title{Takes a virtualspecies object and unlists all of the args in the niche parameters.
If you don't do this after generating random species, ape will barf and die horribly.}
\usage{
unpack.species(species)
}
\arguments{
\item{species}{A virtualspecies object}
}
\value{
A virtualspecies object with its guts unpacked slightly
}
\description{
Takes a virtualspecies object and unlists all of the args in the niche parameters.
If you don't do this after generating random species, ape will barf and die horribly.
}
