% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc.clade.R
\name{vc.clade}
\alias{vc.clade}
\title{Defining a class for vc.clade, which is a container for clades of simulated virtual species and related items.}
\usage{
vc.clade(species = NA, tree = NA, root.species = NA, sim.args = NA)
}
\arguments{
\item{species}{A list of vc.species objects.}

\item{tree}{A tree showing the relationships between the species.}

\item{root.species}{A vc.species object for the root species.}

\item{sim.args}{The arguments used to call evolve.clade for the initial simulation.}
}
\value{
A vc.clade object.
}
\description{
Defining a class for vc.clade, which is a container for clades of simulated virtual species and related items.
}
