% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc.species.R
\name{vc.species}
\alias{vc.species}
\title{Defining a class for vc.species.  Each species gets:}
\usage{
vc.species(
  virtualspecies = NA,
  actual.range = NA,
  presence.points = NA,
  background.points = NA,
  species.name = NA,
  models = NA
)
}
\arguments{
\item{virtualspecies}{A virtualspecies object which contains their niche and true suitability of habitat}

\item{actual.range}{A raster with the actual range they occur in}

\item{presence.points}{A data frame with sampled localities}

\item{background.points}{A data frame with absence/pseudoabsence/background localities}

\item{species.name}{A character vector with the species name}

\item{models}{A list of models that are made for the species, which will be stuffed in there as we go along}
}
\value{
A vc.species object
}
\description{
Defining a class for vc.species.  Each species gets:
}
