% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allopatrify.R
\name{allopatrify}
\alias{allopatrify}
\title{Takes a vc.species object, a raster template, and a buffer width.  Breaks combined range apart into N clusters using kmeans clustering, and then assigns each cluster to the species whose PA centroid is closest to that cluster. Then it buffers the intersection of the cluster and the species PA out by buffer.width, and multiplies by the suitability raster.  That gives us the suitability of habitat within the actual range of the species.  Everything is then stuffed into a vc.clade object.}
\usage{
allopatrify(
  x,
  buffer.width = 1,
  plot = TRUE,
  split.cols = c("lon", "lat"),
  beta = 0.5,
  alpha = -7e-05,
  env = NA,
  npres = NA,
  min.suitable = 10,
  nreps = 1000
)
}
\arguments{
\item{x}{A vc.clade object to allopatrify}

\item{buffer.width}{Controls the post-allpatrification expansion of species' ranges.  Higher numbers create more range overlap, units are decimal degrees.}

\item{plot}{Controls whether to print plots when generating PA data}

\item{split.cols}{Vector of column names to use for partitioning species occurrences.}

\item{beta}{Beta parameter past to convertToPA from virtual species}

\item{alpha}{Alpha parameter past to convertToPA from virtual species}

\item{env}{Optional stack of environmental layers, which will eventually be used to partition species based on environment as well as geography.}

\item{npres}{Number of presence points per species to use for partitioning ranges.}

\item{min.suitable}{Minimum number of "suitable" grid cells in a species range}

\item{nreps}{Number of replicates to try for permuting ranges to match phylogeny}
}
\value{
A vc.clade object with ranges added to the "actual range" variable for each species
}
\description{
Takes a vc.species object, a raster template, and a buffer width.  Breaks combined range apart into N clusters using kmeans clustering, and then assigns each cluster to the species whose PA centroid is closest to that cluster. Then it buffers the intersection of the cluster and the species PA out by buffer.width, and multiplies by the suitability raster.  That gives us the suitability of habitat within the actual range of the species.  Everything is then stuffed into a vc.clade object.
}
